Consolidate framework classes
