--- src/KiBoards/Framework/ITestContextMessage.cs
@@ -1,9 +1,0 @@
-﻿using Xunit.Abstractions;
-
-namespace KiBoards.Framework
-{
-    internal interface ITestContextMessage : IMessageSinkMessage
-    {
-        public object Context { get; set; }
-    }
-}
