--- src/KiBoards/Framework/TestContextMessage.cs
@@ -1,14 +1,0 @@
-﻿using Xunit.Abstractions;
-
-namespace KiBoards.Framework
-{
-    internal class TestContextMessage : ITestContextMessage
-    {
-        public object Context { get; set; }
-
-        public TestContextMessage(object context)
-        {
-            Context = context;
-        }
-    }
-}
